requirements.txt
# Django framework
Django==3.2.20

# Django REST framework para API y JWT
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

# SQLite3 ya está incluido en la instalación estándar de Python, no es necesario agregarlo aquí.

# Bibliotecas para scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3  # Usado para mejorar el análisis de HTML con BeautifulSoup

# Machine Learning y procesamiento de texto
scikit-learn==1.3.0  # Para TfidfVectorizer y cosine_similarity
numpy==1.25.0  # Dependencia común para scikit-learn
Faker==19.2.0  # Para generar datos de prueba

# Manipulación de XML
xmltodict==0.13.0  # Alternativa ligera para el manejo de XML

# Algoritmo de Levenshtein para comparación de cadenas
python-Levenshtein==0.21.0  # Utilizado para cálculos de distancia de edición

# Generación de respuestas JSON
# (No es necesario instalar `JsonResponse`, es parte de Django)

# Manipulación y creación de usuarios en Django
# (El modelo User ya está incluido en Django, no se necesita un paquete adicional)

